#117 fix for changed uri path
