+++ docker/docker-compose_dev.yaml
@@ -49,6 +49,7 @@ services:
       #      - TOOLTEMPLATE=${TOOLTEMPLATE}
       - TOOLPATH=${TOOLPATH}
       - TOOLREPO=${TOOLREPO}
+      - URIVERSION=${URIVERSION}
     labels:
       - "traefik.enable=true"
       - "traefik.http.routers.`tennant${GC_BASE}`-server.entrypoints=http"
